+++ frontend/src/router/Router.tsx
@@ -1,11 +1,4 @@
-import { createContext, useContext, useState } from 'react';
-import {
-  Navigate,
-  Route,
-  Routes,
-  useLocation,
-  useNavigate,
-} from 'react-router-dom';
+import { Navigate, Route, Routes, useLocation } from 'react-router-dom';
 import { ChatRoomList } from 'components/pages/ChatRoomList';
 import { DirectMessage } from 'components/pages/DirectMessage';
 import { Game } from 'components/pages/Game';
@@ -13,63 +6,9 @@ import { GameSelect } from 'components/pages/GameSelect';
 import { UserList } from 'components/pages/UserList';
 import { HeaderLayout } from 'components/templates/HeaderLayout';
 import { Login } from '../components/pages/Login';
+import { LoginPage } from '../components/pages/LoginPage';
 import { Page404 } from '../components/pages/Page404';
-
-type AuthContextType = {
-  user: string;
-  signin: (user: string, callback: VoidFunction) => void;
-  signout: (callback: VoidFunction) => void;
-};
-
-// あまり良くない書き方とのことなので、後ほど検討(参考：https://qiita.com/johnmackay150/items/88654e5064290c24a32a)
-// const AuthContext = createContext<AuthContextType | undefined>(undefined);
-const AuthContext = createContext<AuthContextType>({
-  user: '',
-  signin: (newUser: string, callback: VoidFunction) => {
-    callback();
-  },
-  signout: (callback: VoidFunction) => {
-    callback();
-  },
-});
-
-const fakeAuthProvider = {
-  isAuthenticated: false,
-  signin(callback: VoidFunction) {
-    fakeAuthProvider.isAuthenticated = true;
-    setTimeout(callback, 100); // fake async
-  },
-  signout(callback: VoidFunction) {
-    fakeAuthProvider.isAuthenticated = false;
-    setTimeout(callback, 100);
-  },
-};
-
-const AuthProvider = ({ children }: { children: React.ReactNode }) => {
-  const [user, setUser] = useState('');
-
-  const signin = (newUser: string, callback: VoidFunction) => {
-    return fakeAuthProvider.signin(() => {
-      setUser(newUser);
-      callback();
-    });
-  };
-
-  const signout = (callback: VoidFunction) => {
-    return fakeAuthProvider.signout(() => {
-      setUser('');
-      callback();
-    });
-  };
-
-  const value = { user, signin, signout };
-
-  return <AuthContext.Provider value={value}>{children}</AuthContext.Provider>;
-};
-
-const useAuth = () => {
-  return useContext(AuthContext);
-};
+import { AuthProvider, useAuth } from './providers/useAuthProvider';
 
 const RequireAuth = ({ children }: { children: JSX.Element }) => {
   const auth = useAuth();
@@ -80,38 +19,6 @@ const RequireAuth = ({ children }: { children: JSX.Element }) => {
   }
 
   return children;
-};
-
-const LoginPage = () => {
-  const navigate = useNavigate();
-  const auth = useAuth();
-
-  // 修正方法わからないため一旦eslint回避
-  // ここの'/user-list'はチュートリアルでは'/'だったけど'/user-list'にすると想定通りの動きになった
-  // eslint-disable-next-line @typescript-eslint/no-unsafe-member-access
-  const to = '/user-list';
-
-  function handleSubmit(event: React.FormEvent<HTMLFormElement>) {
-    event.preventDefault();
-
-    const formData = new FormData(event.currentTarget);
-    const username = formData.get('username') as string;
-
-    auth.signin(username, () => {
-      navigate(to as string, { replace: true });
-    });
-  }
-
-  return (
-    <div>
-      <form onSubmit={handleSubmit}>
-        <label>
-          Username: <input name="username" type="text" />
-        </label>{' '}
-        <button type="submit">Login</button>
-      </form>
-    </div>
-  );
 };
 
 export const Router = (): React.ReactElement | null => {
